pyproject.toml
[tool.poetry]
name = "limesurvey_etl"
version = "0.1.0"
description = ""
authors = ["Timo Hartmann <timo.hartmann@posteo.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.0"
pydantic = "1.10.9"
pandas = "1.5.1"
SQLAlchemy = "1.4.42"
PyMySQL = "1.0.2"
pyyaml = "6.0"
python-dotenv = "1.0.0"
sshtunnel = "0.4.0"
psycopg2 = "^2.9.9"

[tool.poetry.group.dev.dependencies]
mkdocs = "1.5.*"
mkdocstrings-python = "1.7.*"
mkdocs-material = "9.*"
mkdocs-table-reader-plugin = "2.0.*"
pytest = "7.4.*"
testcontainers = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
